backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
supabase==2.7.4
google-generativeai==0.3.2
python-multipart==0.0.9
httpx<0.26,>=0.24
redis==5.0.1
celery==5.3.6
